backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.11.3
PyJWT==2.10.1
redis
